# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Solo D&D 5e Companion App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{campaignId}/quests:
    get:
      operationId: listQuests
      tags: [quests]
      summary: Get quest log for a campaign
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quest log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quest"
    post:
      operationId: createQuest
      tags: [quests]
      summary: Add a quest to the log
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestBody"
      responses:
        "201":
          description: Created quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /campaigns/{campaignId}/quests/{questId}:
    put:
      operationId: updateQuest
      tags: [quests]
      summary: Update a quest
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
        - name: questId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestBody"
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
...
components:
  schemas:
...
    Quest:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, failed]
        isMain:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, campaignId, title, status, isMain, createdAt, updatedAt]
    CreateQuestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, failed]
        isMain:
          type: boolean
      required: [title, status, isMain]
    UpdateQuestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, failed]
        isMain:
          type: boolean
...